:art: Cleanup yaml command creation
