+++ pkg/cli/glazed_layer.go
@@ -1,0 +1,276 @@
+package cli
+
+import (
+	_ "embed"
+	"github.com/go-go-golems/glazed/pkg/cmds"
+	"github.com/go-go-golems/glazed/pkg/cmds/layers"
+	"github.com/go-go-golems/glazed/pkg/cmds/parameters"
+	"github.com/go-go-golems/glazed/pkg/formatters"
+	"github.com/go-go-golems/glazed/pkg/middlewares"
+	"github.com/pkg/errors"
+	"github.com/spf13/cobra"
+)
+
+// Helpers for cobra commands
+
+type GlazedParameterLayers struct {
+	FieldsFiltersParameterLayer *FieldsFiltersParameterLayer
+	OutputParameterLayer        *OutputParameterLayer
+	RenameParameterLayer        *RenameParameterLayer
+	ReplaceParameterLayer       *ReplaceParameterLayer
+	SelectParameterLayer        *SelectParameterLayer
+	TemplateParameterLayer      *TemplateParameterLayer
+}
+
+func (g *GlazedParameterLayers) AddFlag(*parameters.ParameterDefinition) {
+	panic("not supported me")
+}
+
+func (g *GlazedParameterLayers) GetParameterDefinitions() map[string]*parameters.ParameterDefinition {
+	ret := make(map[string]*parameters.ParameterDefinition)
+	for k, v := range g.RenameParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	for k, v := range g.OutputParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	for k, v := range g.SelectParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	for k, v := range g.TemplateParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	for k, v := range g.FieldsFiltersParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	for k, v := range g.ReplaceParameterLayer.GetParameterDefinitions() {
+		ret[k] = v
+	}
+
+	return ret
+}
+
+func (g *GlazedParameterLayers) InitializeStructFromDefaults(interface{}) error {
+	panic("implement me")
+}
+
+func (g *GlazedParameterLayers) AddFlagsToCobraCommand(cmd *cobra.Command, defaults interface{}) error {
+	err := g.OutputParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+	err = g.SelectParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+	err = g.RenameParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+	err = g.TemplateParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+	err = g.FieldsFiltersParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+	err = g.ReplaceParameterLayer.AddFlagsToCobraCommand(cmd, defaults)
+	if err != nil {
+		return err
+	}
+
+	layers.SetFlagGroupOrder(cmd, []string{
+		"glazed-output",
+		"glazed-select",
+		"glazed-template",
+		"glazed-fields-filter",
+		"glazed-rename",
+		"glazed-replace",
+	})
+
+	return nil
+}
+
+func (g *GlazedParameterLayers) ParseFlagsFromCobraCommand(cmd *cobra.Command) (map[string]interface{}, error) {
+	ps, err := g.OutputParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	ps_, err := g.SelectParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	for k, v := range ps_ {
+		ps[k] = v
+	}
+	ps_, err = g.RenameParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	for k, v := range ps_ {
+		ps[k] = v
+	}
+	ps_, err = g.TemplateParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	for k, v := range ps_ {
+		ps[k] = v
+	}
+	ps_, err = g.FieldsFiltersParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	for k, v := range ps_ {
+		ps[k] = v
+	}
+	ps_, err = g.ReplaceParameterLayer.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, err
+	}
+	for k, v := range ps_ {
+		ps[k] = v
+	}
+	return ps, nil
+}
+
+func NewGlazedParameterLayers() (*GlazedParameterLayers, error) {
+	fieldsFiltersParameterLayer, err := NewFieldsFiltersParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	outputParameterLayer, err := NewOutputParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	renameParameterLayer, err := NewRenameParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	replaceParameterLayer, err := NewReplaceParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	selectParameterLayer, err := NewSelectParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	templateParameterLayer, err := NewTemplateParameterLayer()
+	if err != nil {
+		return nil, err
+	}
+	return &GlazedParameterLayers{
+		FieldsFiltersParameterLayer: fieldsFiltersParameterLayer,
+		OutputParameterLayer:        outputParameterLayer,
+		RenameParameterLayer:        renameParameterLayer,
+		ReplaceParameterLayer:       replaceParameterLayer,
+		SelectParameterLayer:        selectParameterLayer,
+		TemplateParameterLayer:      templateParameterLayer,
+	}, nil
+}
+
+func SetupProcessor(ps map[string]interface{}) (
+	*cmds.GlazeProcessor,
+	formatters.OutputFormatter,
+	error,
+) {
+	templateSettings, err := NewTemplateSettings(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+	outputSettings, err := NewOutputFormatterSettings(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+	selectSettings, err := NewSelectSettingsFromParameters(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+	renameSettings, err := NewRenameSettingsFromParameters(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+	fieldsFilterSettings, err := NewFieldsFilterSettings(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+	replaceSettings, err := NewReplaceSettingsFromParameters(ps)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	outputSettings.UpdateWithSelectSettings(selectSettings)
+	fieldsFilterSettings.UpdateWithSelectSettings(selectSettings)
+	templateSettings.UpdateWithSelectSettings(selectSettings)
+
+	of, err := outputSettings.CreateOutputFormatter()
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "Error creating output formatter")
+	}
+
+	// rename middlewares run first because they are used to clean up column names
+	// for the following middlewares too.
+	// these following middlewares can create proper column names on their own
+	// when needed
+	err = renameSettings.AddMiddlewares(of)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "Error adding rename middlewares")
+	}
+
+	err = templateSettings.AddMiddlewares(of)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "Error adding template middlewares")
+	}
+
+	if (outputSettings.Output == "json" || outputSettings.Output == "yaml") && outputSettings.FlattenObjects {
+		mw := middlewares.NewFlattenObjectMiddleware()
+		of.AddTableMiddleware(mw)
+	}
+	fieldsFilterSettings.AddMiddlewares(of)
+
+	err = replaceSettings.AddMiddlewares(of)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "Error adding replace middlewares")
+	}
+
+	var middlewares_ []middlewares.ObjectMiddleware
+	if !templateSettings.UseRowTemplates && len(templateSettings.Templates) > 0 {
+		ogtm, err := middlewares.NewObjectGoTemplateMiddleware(templateSettings.Templates)
+		if err != nil {
+			return nil, nil, errors.Wrapf(err, "Could not process template argument")
+		}
+		middlewares_ = append(middlewares_, ogtm)
+	}
+
+	gp := cmds.NewGlazeProcessor(of, middlewares_)
+	return gp, of, nil
+}
+
+// Deprecated: Use SetupProcessor instead, and create a proper glazed.Command for your command.
+// TODO(manuel, 2023-02-21): This is here to facilitate legacy linking, but really all commands
+// using this should move towards implementing a proper glazed.Command.
+// See #150 and for example JsonCmd in glaze
+// lint:ignore
+func CreateProcessorLegacy(cmd *cobra.Command) (
+	*cmds.GlazeProcessor,
+	formatters.OutputFormatter,
+	error,
+) {
+	gpl, err := NewGlazedParameterLayers()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	ps, err := gpl.ParseFlagsFromCobraCommand(cmd)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return SetupProcessor(ps)
+}
